requirements.txt
wtforms~=3.1.2
SQLAlchemy~=2.0.29
bcrypt~=4.1.3
pillow~=10.4.0
flask~=3.0.3